Refactor component structure and update tab label
Moved recipe and pantry related components into feature-specific subfolders for better organization. Updated import paths in affected files. Renamed 'All Houses' tab to 'All Locations' in PantryClient and corresponding test to improve clarity.
